+++ fakeredis/_server.py
@@ -5,7 +5,7 @@ import weakref
 from collections import defaultdict
 from typing import Dict, Tuple, Any, List, Optional, Union
 
-from fakeredis._acl import AccessControlList
+from fakeredis.model import AccessControlList
 from fakeredis._helpers import Database, FakeSelector
 
 LOGGER = logging.getLogger("fakeredis")
