+++ fakeredis/commands_mixins/streams_mixin.py
@@ -5,7 +5,7 @@ import fakeredis._msgs as msgs
 from fakeredis._command_args_parsing import extract_args
 from fakeredis._commands import Key, command, CommandItem, Int
 from fakeredis._helpers import SimpleError, casematch, OK, current_time, Database, SimpleString
-from fakeredis._stream import XStream, StreamRangeTest, StreamGroup, StreamEntryKey
+from fakeredis.model import XStream, StreamRangeTest, StreamGroup, StreamEntryKey
 
 
 class StreamsCommandsMixin:
