+++ fakeredis/model/__init__.py
@@ -1,0 +1,28 @@
+from redis.commands.timeseries import TimeSeries
+
+from ._acl import AccessControlList
+from ._command_info import (
+    get_all_commands_info,
+    get_command_info,
+    get_categories,
+    get_commands_by_category,
+)
+from ._stream import XStream, StreamGroup, StreamEntryKey, StreamRangeTest
+from ._timeseries_model import AGGREGATORS, TimeSeriesRule
+from ._zset import ZSet
+
+__all__ = [
+    "AccessControlList",
+    "XStream",
+    "StreamRangeTest",
+    "StreamGroup",
+    "StreamEntryKey",
+    "ZSet",
+    "TimeSeries",
+    "TimeSeriesRule",
+    "AGGREGATORS",
+    "get_all_commands_info",
+    "get_command_info",
+    "get_categories",
+    "get_commands_by_category",
+]
