+++ fakeredis/model/_stream.py
