+++ fakeredis/model/_timeseries_model.py
@@ -5,7 +5,6 @@ from fakeredis._helpers import Database, SimpleError
 
 
 class TimeSeries:
-
     def __init__(
         self,
         name: bytes,
