+++ fakeredis/stack/_timeseries_mixin.py
@@ -5,7 +5,7 @@ from fakeredis import _msgs as msgs
 from fakeredis._command_args_parsing import extract_args
 from fakeredis._commands import command, Key, CommandItem, Int, Float, Timestamp
 from fakeredis._helpers import Database, SimpleString, OK, SimpleError, casematch
-from ._timeseries_model import TimeSeries, TimeSeriesRule, AGGREGATORS
+from fakeredis.model import TimeSeries, TimeSeriesRule, AGGREGATORS
 
 
 class TimeSeriesCommandsMixin:  # TimeSeries commands
